Tidy up reader code (#12)
I tried to refactor the code so that I didn't have to rely on the hcl code to create the VariableBlock struct, but it didn't work out because of the nested variable block. So instead I'm going to tidy up the code a little here in general, and use a new method to check if a hcl.Expression is generated from a missing field, in `filterMissingExpression`.