Implement more of the CLI (#12)
* Implement some more CLI

Next up is adding slog and slog options. Also adding unit tests for some of the options.

This PR adds `--output`,`--overwrite`, `--stdout` and `--allow-empty` and implements each of them. There are some edge cases I need to cover with logging and the `stdout` option, but otherwise this is a first draft of the 'complete' CLI.

* Move flags comment to somewhere more visible

* Improve CLI general behaviour

cmd.go now prints usage when a flag error occurs, and is structured so that it throws around fewer global variables.