Add `--export-variables` and `--escape-json` (#19)
Add --export-variables and --escape-json

Also move some code to the reader package, since it is needed by the new json package.

This returns a JSON blob of the variables with behaviour described in the README file. I also disabled the default behaviour of Go where it escapes certain characters for HTML reasons because I don't think it will be needed here in the majority of cases, and it makes the condition statements harder to read if you're not using a special app like jq.