Allow mulitple validation blocks (#38)
* Add support for multiple validation blocks per variable

This was missing from the original design because I was unaware that
Terraform allows this. It was pretty straightforward to add though, I
just needed to change the internal `Validation` field to be a slice and
then apply each condition from the slice to the variable.

Mixing and matching rules has not been
tested extensively but hopefully all the JSON schema fields should play
nicely together.

In particular, it is already the case that if multiple
rules specify a minimum or maximum value, then only the "lowest maximum"
and "highest minimum" values are considered, as would be expected.

Certain other fields will not work if used multiple times, such as
`pattern` (regex). Whichever regex is specified last will take
precedence.

Because of edge cases like this, multiple validation rules should be
used with caution.

* Update docs to reflect support for multiple validation blocks