feat: add `ignore-variable` flag (#36)
* feat: add `ignore-variable` flag

This flag allows to ignore variable by name when generating a schema, or exporting variables.

Repeating this flag will allow to ignore multiple variables.

* Ignore the ignored fields in `make test-data`, and fix linting error

This makes sure that when regenerating the test data with `make
test-data`, the output is consistent with the output generated by the
tests, which now expect the fields "ignored" and "also_ignored" to be
skipped.

---------

Co-authored-by: Alexis Couvreur <alexiscouvreur.pro@gmail.com>